refactor(page): extract useMirrorManager hook (phase 3 start)
Move the world-space mirror tool — live reflection preview, off-thread
geometry baking via Web Worker, and session finalize — out of page.tsx
into features/mirror/useMirrorManager.ts, following the existing
manager-hook pattern. The hook takes scene/transformMgr plus the
destructive-modal and persistence-suppression callbacks as deps, and
returns { handleMirror, flushPendingBake, mirrorSessionRef }. page.tsx
keeps the finalizeMirrorSessionRef forward-ref (consumed by the
transform-mode setter, hotkeys, and toolbar) wired to the hook's
flushPendingBake.

page.tsx: 21,949 -> 21,604 lines. Verified via `tsc --noEmit`.
